Split events from main loop, fixes data length for future analysis
